refactor(annotations): :truck: move BindingInfo to annotations package
