+++ pkg/controller/servicebindingrequest/annotations/bindinginfo_test.go
@@ -1,4 +1,4 @@
-package bindinginfo
+package annotations
 
 import (
 	"fmt"
